% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSOM.R
\docType{data}
\name{classSOM}
\alias{classSOM}
\title{Sample Data}
\format{
A data frame with 9757 rows and 4 columns.
}
\source{
example data
}
\usage{
classSOM
}
\description{
This is a sample dataframe included in the package for demonstration.
}
\keyword{datasets}
