% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_displacements.R
\name{create_displacements}
\alias{create_displacements}
\title{Automatic detection of displacement types}
\usage{
create_displacements(model, flip_threshold = -0.85, delay_threshold = 0.85)
}
\arguments{
\item{model}{A trained SOM model.}

\item{flip_threshold}{The threshold for detecting flip displacements.}

\item{delay_threshold}{The threshold for detecting delay displacements.}
}
\value{
A data frame containing the detected displacement types.
}
\description{
This function automatically detects displacement types in a trained SOM model.
}
