% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mappings.R
\name{create_mappings}
\alias{create_mappings}
\title{Create mappings and class assignments}
\usage{
create_mappings(model, train_data, data_list, species_names, classes)
}
\arguments{
\item{model}{A trained SOM model.}

\item{train_data}{Dataframe used for build "model"}

\item{data_list}{A list of matrices or data frames with scale expression data.}

\item{species_names}{Character vector of species names corresponding to the data.}

\item{classes}{A data frame with features class tags.}
}
\value{
A named list containing the mappings, classes, and neuron sizes.
}
\description{
This function creates mappings and class assignments based on a trained
model and input data. The return includes the number of features mapped to
 each neuron.
}
