% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datSOM.R
\docType{data}
\name{datSOM}
\alias{datSOM}
\title{Sample Data}
\format{
A data frame with 9757 rows and 12 columns.
}
\source{
example data
}
\usage{
datSOM
}
\description{
This is a sample dataframe included in the package for demonstration.
}
\keyword{datasets}
