% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_edges.R
\name{net_edges}
\alias{net_edges}
\title{Create network edges}
\usage{
net_edges(
  mappings,
  reference_name,
  test_name,
  add_disp_type = FALSE,
  disp_type = displacements
)
}
\arguments{
\item{mappings}{A named list containing mappings, classes, and neuron sizes.Output from create_mappings}

\item{reference_name}{The name of the species, condition or samples to be used as reference.}

\item{test_name}{The name of the species, condition or samples to be used as test variable.}

\item{add_disp_type}{Logical value indicating whether to add displacement types to the network edges.}

\item{disp_type}{A data frame containing the displacement types. Valid if add_disp_type = TRUE}
}
\value{
A data frame containing the network links and edges. Object used for displacements visualization.
}
\description{
This function creates network edges based on mappings and class assignments.
}
