% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_map_size.R
\name{opt_map_size}
\alias{opt_map_size}
\title{Optimal SOM map size determination}
\usage{
opt_map_size(data, threshold = 0.8, init_dim = 5, max_iterations = 1000)
}
\arguments{
\item{data}{species data for build reference SOM. A data matrix or data frame without NAs.}

\item{threshold}{The neuron correlation threshold for reducing the SOM grid size.}

\item{init_dim}{The initial SOM grid dimensions.}

\item{max_iterations}{The maximum number of iterations for SOM training.}
}
\value{
A list with the resulting SOM grid size, dimensions, and correlation matrix.
}
\description{
This function determines the optimal self-organizing map (SOM) grid size for a given dataset based on neuron correlations.
}
