% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_displacements.R
\name{plot_displacements}
\alias{plot_displacements}
\title{Create a Graph Visualization of Displacements}
\usage{
plot_displacements(
  mappings,
  links,
  reference_species,
  layout = "circle",
  node_order = NULL,
  color_edges,
  color_scale = NULL
)
}
\arguments{
\item{mappings}{A list containing the mappings and neuron sizes. From create_mappings().}

\item{links}{A data frame containing the network edges. From create_net_edges().}

\item{reference_species}{A character vector specifying the reference species.}

\item{layout}{The layout of the graph. From igraph parameter: "circle", "linear", "grid", etc. Default is "circle".}

\item{node_order}{An optional vector specifying the order of nodes in the graph. Default is NULL.}

\item{color_edges}{A character string specifying the column name for edge colors.}

\item{color_scale}{An optional vector specifying the color scale for the edges. Default is NULL.}
}
\value{
A ggplot object representing the graph visualization of displacements.
}
\description{
This function creates a graph visualization of displacements based on the mappings and links data.
}
