% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_condition.R
\name{scale_condition}
\alias{scale_condition}
\title{Scale data}
\usage{
scale_condition(..., rnames, condition_names)
}
\arguments{
\item{...}{Numeric vectors or matrices of condition data.}

\item{rnames}{Character vector of row names for the resulting scaled data.}

\item{condition_names}{Character vector of condition names.}
}
\value{
A named list of scaled data.
}
\description{
This function scales each control or treatment condition data individually by applying a log transformation and
standardizing the columns.
}
