% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_species.R
\name{scale_species}
\alias{scale_species}
\title{Scale data}
\usage{
scale_species(..., rnames, species_names)
}
\arguments{
\item{...}{Numeric vectors or matrices of species data.}

\item{rnames}{Character vector of row names for the resulting scaled data.}

\item{species_names}{Character vector of species names.}
}
\value{
A named list of scaled data.
}
\description{
This function scales the species data individually by applying a log transformation and
standardizing the columns.
}
